+++ src/optional_tasks/times.py
@@ -1,6 +1,10 @@
 import datetime as dt
 import time
 
+SECOND_IN_MINUTE = 60
+SECOND_IN_HOUR = SECOND_IN_MINUTE * 60
+SECOND_IN_DAY = SECOND_IN_HOUR * 24
+
 def datetime_from_utc_to_local(utc_datetime):
     now_timestamp = time.time()
     offset = dt.datetime.fromtimestamp(now_timestamp) - dt.datetime.utcfromtimestamp(now_timestamp)
@@ -20,5 +24,6 @@ def get_current_unix_time() -> float:
 def unix_time_to_iso(unix_time: float):
     return dt.datetime.fromtimestamp(unix_time).strftime("%Y-%m-%d %H:%M:%S")
 
-def get_unix_day_start_time():
-    return get_unix_time(datetime_from_local_to_utc(dt.datetime.combine(dt.date.today(), dt.time(00, 00, 00))))
+def get_unix_day_start_time(unix_time = get_current_unix_time()):
+    datetime = dt.datetime.fromtimestamp(unix_time)
+    return get_unix_time(datetime_from_local_to_utc(dt.datetime.combine(datetime.date(), dt.time(00, 00, 00))))